+++ firmware/application/freqman.cpp
@@ -315,7 +315,6 @@ bool save_freqman_file(std::string& file_stem, freqman_db& db) {
             auto& entry = db[n];
             get_freq_string(entry, item_string);
             freqman_file.write_line(item_string);
-            delete &item_string;
         }
         delete_file(freq_file_path);
         rename_file(tmp_freq_file_path, freq_file_path);
